IE11 Fixes
IE11 doesn't support CSS variables.
IE11 doesn't support webp.
I am doing this because I am trying to use a Surface RT and it can run
only IE11 and I want to be able to see twitter.
There are still some small layout problems, but for now it is enough.